Move metrics tracker from shared utils into plugin reporting
The createMetricsTracker helper previously lived in src/shared/utils/metrics.js even though it is only consumed by plugin reporting and project-index modules. Relocate it to src/plugin/src/reporting/metrics-tracker.js, update imports, and move its tests so the tracker stays within the plugin boundary.
