+++ src/plugin/src/reporting/index.js
@@ -1,1 +1,1 @@
-export * from "../../../shared/metrics.js";
+export * from "./metrics-tracker.js";
